change axios request to injection of Service
