+++ src/modules/signup-timetables/signup-timetables.repository.ts
@@ -19,7 +19,8 @@ export class SignuptimetablesRepository extends Repository<Signuptimetable> {
         time: updateData.datesArray[date] + " " + updateData.time,
         userId: updateData.userId,
       });
-      console.log(updateData.time + " 시간입니다");
+      // console.log("1회성");
+      // console.log(updateData.time + " 시간입니다");
       console.log(updateData.datesArray[date]);
       console.log(timetable);
       delay(0.00000001);
